requirements.txt
streamlit
openai
python-dotenv
PyPDF2
python-docx
fpdf